+++ plugins/animationaddon/src/particle.cpp
@@ -276,8 +276,10 @@ ParticleAnim::updateBB (CompOutput &output)
 	    h += (h * part.h_mod) * part.life;
 
 	    Box particleBox =
-		{part.x - w, part.x + w,
-		 part.y - h, part.y + h};
+	    {
+		static_cast <short int> (part.x - w), static_cast <short int> (part.x + w),
+		static_cast <short int> (part.y - h), static_cast <short int> (part.y + h)
+	    };
 
 	    mAWindow->expandBBWithBox (particleBox);
 	}
